Refactor reservations route to be RestFul
- Delete reservations occurences
- Use begin param of getPlaces
